refactor(tools): move tools src -> ./
