+++ tools/testing/setup.ts
@@ -1,4 +1,4 @@
 import { expect } from 'vitest';
-import vueSnapshotSerializer from './serializer.js';
+import vueSnapshotSerializer from './serializer';
 
 expect.addSnapshotSerializer(vueSnapshotSerializer);
